Add auction

+++ src/TradingApp.Core/Models/UserStock.cs
@@ -6,5 +6,5 @@ public class UserStock
     public int UserId { get; set; }
     public int StockId { get; set; }
     public double TotalPrice { get; set; }
-    public int StockCount { get; set; }
+    public double StockCount { get; set; }
 }
